requirements.txt
asciimatics==1.12.0
dict_deep==4.1.2
PyYAML==5.4.1
